+++ pgn-vizu.cabal
@@ -16,6 +16,7 @@ cabal-version:       >=1.10
 library
   hs-source-dirs:      src
   exposed-modules:     PGN
+                     , Models
   ghc-options:         -Wall -Werror
   build-depends:       base >= 4.7 && < 5
                      , bytestring
