+++ src/PGN.hs
@@ -1,24 +1,9 @@
 module PGN
-    ( Match(..)
-    , Move(..)
-    , Color(..)
-    , PieceMove(..)
-    , Piece(..)
-    , Capture(..)
-    , Disambiguate(..)
-    , Square
-    , File
-    , Rank
-    , Promotion(..)
-    , Check(..)
-    , Glyph(..)
-    , Annotation(..)
-    , Comment
-    , Header(..)
+    ( parseFilePath
     , ParseError
-    , ResultValue(..)
-    , parseFilePath
     ) where
+
+import Models
 
 import qualified Control.Applicative as A ((<|>))
 import Data.Maybe (fromJust, isJust, fromMaybe)
@@ -27,103 +12,6 @@ import Text.Parsec.ByteString (Parser, parseFromFile)
 
 parseFilePath :: FilePath -> IO (Either ParseError Match)
 parseFilePath file = parseFromFile parseMatch file
-
-data Match = Match { matchHeaders :: [Header]
-                   , matchMoves   :: Move
-                   } deriving (Show)
-
-data Header = Event String
-            | Site String
-            | Date String
-            | Round String
-            | WhitePlayer String
-            | BlackPlayer String
-            | Result ResultValue
-            | WhiteElo Int
-            | BlackElo Int
-            | PlyCount String
-            | Variant String
-            | TimeControl String
-            | ECO String
-            | Opening String
-            | Termination String
-            | Annotator String
-            | Other String String
-              deriving (Show)
-
-data Color = White | Black deriving (Show)
-
-data Move = HalfMove { moveNumber      :: Int
-                     , moveColor       :: Color
-                     , movePieceMove   :: PieceMove
-                     , moveCheck       :: Check
-                     , moveAnnotations :: [Annotation]
-                     , moveNext        :: Move
-                     , moveVariants    :: [Move]
-                     }
-          | End ResultValue
-          | VariantEnd
-           deriving (Show)
-
-data PieceMove = ShortCastle
-               | LongCastle
-               | PieceMove Piece Disambiguate Capture Square
-               | PawnMove { pawnStartFile   :: (Maybe File)
-                          , pawnCapture     :: Capture
-                          , pawnDestination :: Square
-                          , pawnPromotion   :: Promotion
-                          }
-                 deriving (Show)
-
-data Piece = Knight | Bishop | Rook | Queen | King deriving (Show)
-type File = Char
-type Rank = Char
-type Square = (File, Rank)
-data Disambiguate = FileDisambiguate File
-                  | RankDisambiguate Rank
-                  | SquareDisambiguate Square
-                  | NoDisambiguate
-                    deriving (Show)
-data Capture = Capture | NoCapture deriving (Show)
-data Check = None | Check | Mate deriving (Show)
-
-data Annotation = GlyphAnnotation Glyph | CommentAnnotation Comment deriving (Show)
-type Annotations = [Annotation]
-newtype Glyph = Glyph Int deriving (Show)
-type Comment = String
-
-mkAnnotations :: Maybe Glyph -> [Comment] -> [Annotation]
-mkAnnotations g cs = mkGlyphAnnotation g ++ mkCommentAnnotations cs
-  where mkGlyphAnnotation Nothing   = []
-        mkGlyphAnnotation (Just g') = [GlyphAnnotation g']
-        mkCommentAnnotations = map CommentAnnotation
-
-data ResultValue = WhiteWins | BlackWins | Draw | Unknown deriving (Show)
-
-readResultValue :: String -> ResultValue
-readResultValue "1-0"     = WhiteWins
-readResultValue "0-1"     = BlackWins
-readResultValue "1/2-1/2" = Draw
-readResultValue _         = Unknown
-
-readHeader :: String -> String -> Header
-readHeader "Event"       = Event
-readHeader "Site"        = Site
-readHeader "Date"        = Date
-readHeader "Round"       = Round
-readHeader "White"       = WhitePlayer
-readHeader "Black"       = BlackPlayer
-readHeader "Result"      = Result . readResultValue
-readHeader "WhiteElo"    = WhiteElo . read
-readHeader "BlackElo"    = BlackElo . read
-readHeader "PlyCount"    = PlyCount
-readHeader "Variant"     = Variant
-readHeader "TimeControl" = TimeControl
-readHeader "ECO"         = ECO
-readHeader "Opening"     = Opening
-readHeader "Termination" = Termination
-readHeader "Annotator"   = Annotator
-readHeader h             = Other h
 
 parseMatch :: Parser Match
 parseMatch = Match <$> parseHeaders <*> parseFirstMove
@@ -224,7 +112,7 @@ parseDisambiguate = triesOr [ SquareDisambiguate <$> parseSquare
                             , RankDisambiguate   <$> parseRank
                             ] NoDisambiguate
 
-parseSquare :: Parser (File, Rank)
+parseSquare :: Parser Square
 parseSquare = (,) <$> parseFile <*> parseRank
 
 parseFile :: Parser File
@@ -236,10 +124,6 @@ parseRank = oneOf "12345678"
 parseCapture :: Parser Capture
 parseCapture = triesOr [ Capture <$ char 'x'
                        ] NoCapture
-
-data Promotion = PromoteTo Piece
-               | NoPromotion
-                 deriving Show
 
 parsePromotion :: Parser Promotion
 parsePromotion = triesOr [ (PromoteTo Queen ) <$ string "=Q"
